+++ src/synhydro/__init__.py
@@ -39,24 +39,24 @@ from synhydro.core.validation import (
 )
 
 # Generators
-from synhydro.methods.generation.nonparametric.kirsch import KirschGenerator
+from synhydro.methods.generation.hybrid.hmm_knn import HMMKNNGenerator
+from synhydro.methods.generation.hybrid.kirsch import KirschGenerator
+from synhydro.methods.generation.hybrid.multisite_phase_randomization import (
+    MultisitePhaseRandomizationGenerator,
+)
+from synhydro.methods.generation.hybrid.phase_randomization import (
+    PhaseRandomizationGenerator,
+)
+from synhydro.methods.generation.hybrid.warm import WARMGenerator
 from synhydro.methods.generation.nonparametric.knn_bootstrap import (
     KNNBootstrapGenerator,
 )
-from synhydro.methods.generation.nonparametric.multisite_phase_randomization import (
-    MultisitePhaseRandomizationGenerator,
-)
-from synhydro.methods.generation.nonparametric.phase_randomization import (
-    PhaseRandomizationGenerator,
-)
 from synhydro.methods.generation.parametric.arfima import ARFIMAGenerator
-from synhydro.methods.generation.parametric.hmm_knn import HMMKNNGenerator
 from synhydro.methods.generation.parametric.thomas_fiering import ThomasFieringGenerator
 from synhydro.methods.generation.parametric.matalas import MatalasGenerator
 from synhydro.methods.generation.parametric.multisite_hmm import MultiSiteHMMGenerator
 from synhydro.methods.generation.parametric.smarta import SMARTAGenerator
 from synhydro.methods.generation.parametric.sparta import SPARTAGenerator
-from synhydro.methods.generation.parametric.warm import WARMGenerator
 
 # Disaggregators
 from synhydro.methods.disaggregation.temporal.nowak import NowakDisaggregator